requirements.txt
torch
librosa
pyyaml
pytorch_lightning